Format, move list_item to base non-data variable
